fix desktop callback
